reorganized dependencies deduplication
